remove: orphan `Hashable` instance for `Map`
`hashable` 1.3.4.0 comes with instances for container types. Hence,
setting this as a lower bound of the dependency (upper bounds are
currently not used by the project), and removing the orphan instance
from the test-suite.

Alternatively, `CPP` could be used to support older versions of
`Hashable` by conditionally providing the orphan instance.